% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGSsounds.R
\docType{package}
\name{CGSsounds}
\alias{CGSsounds-package}
\alias{CGSsounds}
\title{CGSsounds: A package for accessing soundscape samples}
\description{
The CGSsounds package provides examples of wave objects extracted
from the collection of sounds in the Center for Global Soundscapes,
Purdue University, West Lafayette IN, United States.
}
\references{
For more information see:
\url{https://github.com/panchorivasf/CGSsounds}
}
\keyword{internal}
