% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_reef_pr.R
\docType{data}
\name{coral_reef_pr}
\alias{coral_reef_pr}
\title{Recording of Coral Reef Soundscape in Puerto Rico}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
coral_reef_pr
}
\description{
Recording of Coral Reef Soundscape in Puerto Rico
}
\details{
Recorder: ?
Date: 2017-07-04
Time: 22:00

Sampling rate: 48 kHz.
Duration: 60 seconds.
}
\examples{
\dontrun{
data(coral_reef_pr)
summary(coral_reef_pr)
seewave::spectro(coral_reef_pr)
}

}
\keyword{datasets}
