% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawn_chorus_tanzania.R
\docType{data}
\name{dawn_chorus_tanzania}
\alias{dawn_chorus_tanzania}
\title{Recording of Dawn Chorus in Tanzania}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
dawn_chorus_tanzania
}
\description{
The sounds of insects and birds during the dawn chorus in Issa Valley,
Katavi Region in Western Tanzania.
}
\details{
Date: 2017-03-25
Time: 07:00

Sampling rate: 48 kHz.
Duration: 60 seconds.
}
\examples{
\dontrun{
data(dawn_chorus_tanzania)
summary(dawn_chorus_tanzania)
seewave::spectro(dawn_chorus_tanzania)
}

}
\keyword{datasets}
