% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insects_borneo.R
\docType{data}
\name{insects_borneo}
\alias{insects_borneo}
\title{Recording of Insects from Borneo Island}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
insects_borneo
}
\description{
Recording of insect chorus at midnight in Borneo, Brunei.
}
\details{
Recorder: SongMeter4 (Wildlife Acoustics)
Date: 2014-02-20
Time: 00:00

Sampling rate: 44.1 kHz.
Duration: 60 seconds.
}
\examples{
\dontrun{
data(insects_borneo)
summary(insects_borneo)
seewave::spectro(insects_borneo)
}

}
\keyword{datasets}
