% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insects_nebraska.R
\docType{data}
\name{insects_nebraska}
\alias{insects_nebraska}
\title{Recording of Insects from Nebraska, USA}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
insects_nebraska
}
\description{
Recording of insect chorus at midnight in Nebraska, USA.
}
\details{
Recorder: SongMeter4 (Wildlife Acoustics)
Date: 2017-08-18
Time: 00:00

Sampling rate: 48 kHz.
Duration: 60 seconds.
}
\examples{
\dontrun{
data(insects_nebraska)
summary(insects_nebraska)
seewave::spectro(insects_nebraska)
}

}
\keyword{datasets}
