% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insects_tropical.R
\docType{data}
\name{insects_tropical}
\alias{insects_tropical}
\title{Recording of Tropical Rainforest}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
insects_tropical
}
\description{
The sounds of insects in a tropical rainforest at midnight, recorded
on July 28th, 2016 with a SongMeter recorder (Wildlife Acoustics).
}
\details{
Sampling rate: 44.1 kHz.

Duration: 60 seconds.
}
\examples{
\dontrun{
data(insects_tropical)
summary(insects_tropical)
seewave::spectro(insects_tropical)
}

}
\keyword{datasets}
