% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins_patagonia.R
\docType{data}
\name{penguins_patagonia}
\alias{penguins_patagonia}
\title{Recording of Penguins in Patagonia}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
penguins_patagonia
}
\description{
Recording of a Magellanic Penguin colony in Tierra del Fuego, Patagonia.
}
\details{
Recorder: SongMeter3 (Wildlife Acoustics)
Date: 2016-02-25
Time: 06:00

Sampling rate: 48 kHz.
Duration: 60 seconds.
}
\examples{
\dontrun{
data(penguins_patagonia)
summary(penguins_patagonia)
seewave::spectro(penguins_patagonia)
}

}
\keyword{datasets}
