% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siskin_patagonia.R
\docType{data}
\name{siskin_patagonia}
\alias{siskin_patagonia}
\title{Recording of a Siskin (bird)}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\usage{
siskin_patagonia
}
\description{
Recording of a Black-chinned Siskin (\textit{Spinus barbatus})
song in Tierra del Fuego, Patagonia.
}
\details{
Recorder: SongMeter3 (Wildlife Acoustics)
Date: 2016-02-18
Time: 18:00

Sampling rate: 48 kHz.
Duration: 60 seconds.
}
\examples{
\dontrun{
data(siskin_patagonia)
summary(siskin_patagonia)
seewave::spectro(siskin_patagonia)
}

}
\keyword{datasets}
