% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swamp_tanzania.R
\docType{data}
\name{swamp_tanzania}
\alias{swamp_tanzania}
\title{Recording of a Swamp Soundscape in Western Tanzania}
\format{
An object of class \code{\link[tuneR]{Wave}}.
}
\source{
Original recording available at Xeno-canto: https://xeno-canto.org/753101.
}
\usage{
swamp_tanzania
}
\description{
A soundscape of insects and birds in a swamp in the Katavi Region,
Tanzania, recorded on 25th of July, 2017, with a SongMeter (Wildlife
Acoustics).
}
\details{
Sampling rate: 48 kHz.

Duration: 60 seconds.
}
\examples{
\dontrun{
data(swamp_tanzania)
summary(swamp_tanzania)
spectro(swamp_tanzania)
}

}
\keyword{datasets}
